Templatelib fixes
