+++ alkos/kernel/test/array_tests.cpp
@@ -1,0 +1,10 @@
+#include <extensions/array.hpp>
+#include <test_module/test.hpp>
+
+class ArrayTest : public TestGroupBase
+{
+};
+
+/**
+ * TODO: Implement tests for the functional module
+ */
