+++ alkos/libc/include/extensions/array.hpp
@@ -1,0 +1,10 @@
+#ifndef LIBC_INCLUDE_EXTENSIONS_ARRAY_HPP_
+#define LIBC_INCLUDE_EXTENSIONS_ARRAY_HPP_
+
+namespace std
+{
+template <>
+
+}
+
+#endif  // LIBC_INCLUDE_EXTENSIONS_ARRAY_HPP_
