+++ alkos/libc/include/extensions/template_lib.hpp
@@ -8,10 +8,27 @@
 #include <extensions/types.hpp>
 #include <extensions/utility.hpp>
 
+#include <extensions/internal/tuple_base.hpp>
+
 #include <assert.h>
 
 namespace TemplateLib
 {
+// ------------------------------
+// NoCopy
+// ------------------------------
+
+struct NoCopy {
+    NoCopy()  = default;
+    ~NoCopy() = default;
+
+    NoCopy(const NoCopy &) = delete;
+    NoCopy(NoCopy &&)      = delete;
+
+    NoCopy &operator=(const NoCopy &) = delete;
+    NoCopy &operator=(NoCopy &&)      = delete;
+};
+
 // ------------------------------
 // Rolled Switch
 // ------------------------------
@@ -124,56 +141,56 @@ constexpr bool HasDuplicateTypes()
 // Type List
 // ------------------------------
 
-template <size_t N, class T, class... Ts>
-struct TypeList {
-    static constexpr size_t size = sizeof...(Ts) + 1;
-
-    static_assert(N < size, "Index out of range");
-    using type = typename TypeList<N - 1, Ts...>::type;
-};
-
-template <class T, class... Ts>
-struct TypeList<0, T, Ts...> {
-    static constexpr size_t size = sizeof...(Ts) + 1;
-
-    using type = T;
-};
-
-template <size_t N, template <size_t> class TypeList>
-struct IterateTypeList {
-    template <class Callable, class... Args>
-    FORCE_INLINE_F static constexpr void Apply(Callable &&func, Args &&...args)
-    {
-        func.template operator()<N, typename TypeList<N>::type>(args...);
-        IterateTypeList<N - 1, TypeList>::Apply(
-            std::forward<Callable>(func), std::forward<Args>(args)...
-        );
-    }
-};
-
-template <template <size_t> class TypeList>
-struct IterateTypeList<0, TypeList> {
-    template <class Callable>
-    FAST_CALL constexpr void Apply(Callable &&func)
-    {
-        func.template operator()<0, typename TypeList<0>::type>();
-    }
-};
-
 template <class... Args>
-struct IterateTypes {
-    static_assert(sizeof...(Args) > 0, "Type list must not be empty");
+struct TypeList {
+    // ------------------------------
+    // iterators
+    // ------------------------------
+
+    template <size_t N, class T, class... Ts>
+    struct TypeListIter {
+        static constexpr size_t kSize = sizeof...(Ts) + 1;
+
+        static_assert(N < kSize, "Index out of range");
+        using type = typename TypeListIter<N - 1, Ts...>::type;
+    };
+
+    template <class T, class... Ts>
+    struct TypeListIter<0, T, Ts...> {
+        static constexpr size_t kSize = sizeof...(Ts) + 1;
+
+        using type = T;
+    };
+
+    // ------------------------------
+    // Invokers
+    // ------------------------------
+
+    template <size_t N, class Callable, class... ApplyArgs>
+    FORCE_INLINE_F static constexpr void Apply(Callable &&func, ApplyArgs &&...args)
+    {
+        func.template operator()<N, typename Iterator<N>::type>(args...);
+
+        if constexpr (N > 0) {
+            Apply<N - 1>(std::forward<Callable>(func), std::forward<ApplyArgs>(args)...);
+        }
+    }
+
+    template <class Callable, class... ApplyArgs>
+    FORCE_INLINE_F static constexpr void Apply(Callable &&func, ApplyArgs &&...args)
+    {
+        Apply<kSize - 1>(std::forward<Callable>(func), std::forward<ApplyArgs>(args)...);
+    }
+
+    // ------------------------------
+    // Fields
+    // ------------------------------
+
+    static constexpr size_t kSize = sizeof...(Args);
 
     template <size_t N>
-    using TypeList = TypeList<N, Args...>;
-
-    template <class Callable, class... CallArgs>
-    FAST_CALL constexpr void Apply(Callable &&func, CallArgs &&...args)
-    {
-        IterateTypeList<sizeof...(Args) - 1, TypeList>::Apply(
-            std::forward<Callable>(func), std::forward<CallArgs>(args)...
-        );
-    }
+    using Iterator = TypeListIter<N, Args...>;
+    using Tuple    = std::tuple<Args...>;
 };
 
 template <class T, class... Args>
@@ -182,7 +199,7 @@ NODISCARD FAST_CALL constexpr size_t GetTypeIndexInTypes()
     static_assert(HasTypeOnce<T, Args...>(), "Type must occur exactly once in the tuple");
     size_t idx{};
 
-    IterateTypes<Args...>::Apply([&]<size_t Index, class U>() {
+    TypeList<Args...>::Apply([&]<size_t Index, class U>() {
         if constexpr (std::is_same_v<T, U>) {
             idx = Index;
         }
@@ -191,22 +208,27 @@ NODISCARD FAST_CALL constexpr size_t GetTypeIndexInTypes()
     return idx;
 }
 
+template <class T>
+struct IsTypeList : std::false_type {
+};
+
+template <class... Args>
+struct IsTypeList<TypeList<Args...>> : std::true_type {
+};
+
+template <class T>
+constexpr bool IsTypeList_v = IsTypeList<T>::value;
+
 // ------------------------------
 // Static singleton
 // ------------------------------
 
-class StaticSingletonHelper
+class StaticSingletonHelper : public NoCopy
 {
-    public:
-    StaticSingletonHelper(StaticSingletonHelper const &)            = delete;
-    StaticSingletonHelper &operator=(StaticSingletonHelper const &) = delete;
-
-    StaticSingletonHelper(StaticSingletonHelper &&)            = delete;
-    StaticSingletonHelper &operator=(StaticSingletonHelper &&) = delete;
-
     protected:
     /* Non instantiable */
-    StaticSingletonHelper() = default;
+    StaticSingletonHelper()  = default;
+    ~StaticSingletonHelper() = default;
 };
 
 template <class T>
@@ -271,6 +293,59 @@ bool StaticSingleton<T>::is_instance_inited_ = false;
 template <typename T>
     requires DerivedFromHelper<T>
 unsigned char StaticSingleton<T>::instance_memory_[sizeof(T)]{};
+
+// ------------------------------
+// Settings
+// ------------------------------
+
+template <class TypeListT>
+    requires IsTypeList_v<TypeListT>
+class Settings : public NoCopy
+{
+    // ------------------------------
+    // internals
+    // ------------------------------
+
+    public:
+    using EvenT  = void (*)();
+    using TupleT = TypeListT::Tuple;
+
+    private:
+    struct NodeT {
+        EvenT event;
+        NodeT *next;
+    };
+
+    static constexpr size_t kMaxEvents = 256;
+
+    public:
+    // ------------------------------
+    // Class creation
+    // ------------------------------
+
+    // constexpr Settings(TupleT&& settings, )
+
+    // ------------------------------
+    // Class methods
+    // ------------------------------
+
+    //    template<class T>
+    //    requires std::is_convertible_v<T, size_t>
+    //    FORCE_INLINE_F static constexpr auto GetSetting(const T t_item_idx) {
+    //        const size_t idx = static_cast<size_t>(t_item_idx);
+    //        const
+    //    }
+
+    // ------------------------------
+    // Class fields
+    // ------------------------------
+
+    private:
+    TupleT settings_;
+    NodeT *events_[TypeListT::kSize]{};
+    NodeT events_mem_[kMaxEvents];
+    size_t events_mem_top_{};
+};
 
 }  // namespace TemplateLib
 #endif  // LIBC_INCLUDE_TEMPLATE_LIB_HPP_
